requirements.txt
fastapi
uvicorn
python-multipart
openai
python-docx
pillow
jinja2
fastapi
pydantic
ruff
reportlab
openpyxl
PyPDF2
python-docx
